Moved AreaRangeSeries to folder
